Inferno/requirements.txt
openai>=1.0.0
httpx>=0.24.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
litellm

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
